# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product Management API",
        "description": "RESTful API for products, categories, authentication, image uploads, and soft deletes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Use Bearer Token from login endpoint"
            }
        }
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": ["Auth"],
                "summary": "Register a new user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name","email","password","password_confirmation"],
                                "properties": {
                                    "name": {"type": "string", "example": "Ali"},
                                    "email": {"type": "string", "format": "email", "example": "ali@example.com"},
                                    "password": {"type": "string", "example": "password123"},
                                    "password_confirmation": {"type": "string", "example": "password123"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {"description": "User registered successfully"}
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": ["Auth"],
                "summary": "Login user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["email","password"],
                                "properties": {
                                    "email": {"type": "string", "format": "email", "example": "ali@example.com"},
                                    "password": {"type": "string", "example": "password123"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"description": "Login successful"},
                    "422": {"description": "Invalid credentials"}
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": ["Auth"],
                "summary": "Get authenticated user",
                "security": [{"sanctum": []}],
                "responses": {
                    "200": {"description": "Authenticated user fetched successfully"},
                    "401": {"description": "Unauthenticated"}
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": ["Auth"],
                "summary": "Logout current user",
                "security": [{"sanctum": []}],
                "responses": {
                    "200": {"description": "Logged out successfully"},
                    "401": {"description": "Unauthenticated"}
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": ["Products"],
                "summary": "List products",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "search", "in": "query", "description": "Search by name or description", "schema": {"type": "string"}},
                    {"name": "status", "in": "query", "description": "Filter by status", "schema": {"type": "string", "enum": ["active","inactive"]}},
                    {"name": "category_id", "in": "query", "description": "Filter by category", "schema": {"type": "integer"}},
                    {"name": "min_stock", "in": "query", "description": "Minimum stock", "schema": {"type": "integer"}},
                    {"name": "min_price", "in": "query", "description": "Minimum price", "schema": {"type": "number", "format": "float"}},
                    {"name": "max_price", "in": "query", "description": "Maximum price", "schema": {"type": "number", "format": "float"}},
                    {"name": "sort_by", "in": "query", "description": "Sort field", "schema": {"type": "string"}},
                    {"name": "sort_direction", "in": "query", "description": "Sort direction", "schema": {"type": "string", "enum": ["asc","desc"]}},
                    {"name": "per_page", "in": "query", "description": "Items per page", "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Products fetched successfully"},
                    "401": {"description": "Unauthenticated"}
                }
            },
            "post": {
                "tags": ["Products"],
                "summary": "Create product",
                "security": [{"sanctum": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": ["name","price","stock","status"],
                                "properties": {
                                    "name": {"type": "string", "example": "Laptop"},
                                    "description": {"type": "string", "example": "Gaming laptop"},
                                    "price": {"type": "number", "format": "float", "example": 1200.50},
                                    "stock": {"type": "integer", "example": 5},
                                    "status": {"type": "string", "enum": ["active","inactive"]},
                                    "category_id": {"type": "integer", "example": 1},
                                    "image": {"type": "string", "format": "binary"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {"description": "Product created successfully"},
                    "422": {"description": "Validation error"}
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": ["Products"],
                "summary": "Get single product",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "product", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Product fetched successfully"},
                    "403": {"description": "Forbidden"},
                    "404": {"description": "Not found"}
                }
            },
            "post": {
                "tags": ["Products"],
                "summary": "Update product",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "product", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {"type": "string"},
                                    "description": {"type": "string"},
                                    "price": {"type": "number", "format": "float"},
                                    "stock": {"type": "integer"},
                                    "status": {"type": "string", "enum": ["active","inactive"]},
                                    "category_id": {"type": "integer"},
                                    "image": {"type": "string", "format": "binary"},
                                    "_method": {"type": "string", "example": "PUT"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"description": "Product updated successfully"}
                }
            },
            "delete": {
                "tags": ["Products"],
                "summary": "Soft delete product",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "product", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Product moved to trash successfully"}
                }
            }
        },
        "/api/products/trash": {
            "get": {
                "tags": ["Products"],
                "summary": "List trashed products",
                "security": [{"sanctum": []}],
                "responses": {
                    "200": {"description": "Trashed products fetched successfully"}
                }
            }
        },
        "/api/products/{product}/restore": {
            "post": {
                "tags": ["Products"],
                "summary": "Restore trashed product",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "product", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Product restored successfully"}
                }
            }
        },
        "/api/products/{product}/force": {
            "delete": {
                "tags": ["Products"],
                "summary": "Force delete trashed product",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "product", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Product permanently deleted successfully"}
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": ["Categories"],
                "summary": "List categories",
                "security": [{"sanctum": []}],
                "responses": {
                    "200": {"description": "Categories fetched successfully"}
                }
            },
            "post": {
                "tags": ["Categories"],
                "summary": "Create category",
                "security": [{"sanctum": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["name"],
                                "properties": {
                                    "name": {"type": "string", "example": "Electronics"},
                                    "description": {"type": "string", "example": "Electronic devices"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {"description": "Category created successfully"}
                }
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": ["Categories"],
                "summary": "Get single category",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "category", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Category fetched successfully"},
                    "404": {"description": "Not found"}
                }
            },
            "put": {
                "tags": ["Categories"],
                "summary": "Update category",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "category", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {"type": "string"},
                                    "description": {"type": "string"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {"description": "Category updated successfully"}
                }
            },
            "delete": {
                "tags": ["Categories"],
                "summary": "Delete category",
                "security": [{"sanctum": []}],
                "parameters": [
                    {"name": "category", "in": "path", "required": true, "schema": {"type": "integer"}}
                ],
                "responses": {
                    "200": {"description": "Category deleted successfully"}
                }
            }
        }
    }
}